+++ packages/adminui-vue/components/UTableEntity/components/UTableEntityRoot.vue
@@ -477,7 +477,7 @@
 
 <script>
 const { mapState, mapGetters, mapMutations, mapActions } = require('vuex')
-const TypeProvider = require('../type-provider')
+const ColumnTemplateProvider = require('../column-template-provider')
 
 export default {
   name: 'UTableEntityRoot',
@@ -646,10 +646,8 @@ export default {
     getCellTemplate (column) {
       if (typeof column.template === 'function') {
         return column.template()
-      } else {
-        const dataType = column.attribute?.dataType
-        return TypeProvider.get(dataType).template
       }
+      return ColumnTemplateProvider.getByColumnAttribute(column.attribute).template
     },
 
     showContextMenu ({ event, row, column }) {