+++ packages/adminui-vue/CHANGELOG.md
@@ -9,7 +9,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - new ubConfig parameter `application.uiSettings.adminUI.lookupCacheRefreshIntervalSec` - a timeout (in seconds) after which
    pressing a "Refresh" in UTableEntity will refresh a table data AND lookups caches for entities used by current table.
    Default is 0 - do not refresh lookups. *WARNING* - entities with `Session` and `SessionEntity` cache type will not be refreshed.
- - UTableEntity filters: for columns of type Entity and Many added the ability to set a repository for a dropdown list (dropdown filter)
+ - `UTableEntity` filters: for columns of type Entity and Many added the ability to set a repository for a dropdown list (dropdown filter)
+ - `UTableEntity` ability to register custom cell templates for columns globally and use them through the
+   `customSettings.columnTemplate` of the column attribute
+ - **BREAKING** `adminui-vue/components/UTableEntity/type-provider.js` should not be used anymore (removed):
+ ```javascript
+  // instead of
+  const TypeProvider = require('@unitybase/adminui-vue/components/UTableEntity/type-provider.js')
+  TypeProvider.registerType({...})
+  // use
+  const { columnTemplates } = require('@unitybase/adminui-vue')
+  columnTemplates.registerTemplate({...})    
+ ```
 
 ### Changed
  - `UTableEntity`: lookups are loaded in parallel to speed-up table loading